Create a Manager class
This is basically a conversion of the manager_*()
functions into a C++ class.

+++ src/rpc_server.cc
@@ -26,41 +26,39 @@
 #include "log.h"
 #include "manager.h"
 
-static json _rpc_op_disable(const json &args) {
+
+static json _rpc_op_disable(const json &args, Manager &mgr) {
     const auto &label = args[1]["Label"];
-    manager_set_job_enabled(label, false);
+    mgr.overrideJobEnabled(label, false);
     return {{"error", false}};
 }
 
-static json _rpc_op_enable(const json &args) {
+static json _rpc_op_enable(const json &args, Manager &mgr) {
     const auto &label = args[1]["Label"];
-    manager_set_job_enabled(label, true);
+    mgr.overrideJobEnabled(label, true);
     return {{"error", false}};
 }
 
-static json _rpc_op_kill(const json &args) {
+static json _rpc_op_kill(const json &args, Manager &mgr) {
     const std::string &label = args[1]["Label"];
     const std::string &signame_or_num = args[1]["Signal"];
-    auto job = manager_get_job_by_label(label);
-    job->kill(signame_or_num);
-    return {{"error", false}};
-}
-
-static json _rpc_op_list(const json &) {
-    // FIXME: handle Label argument
-    return manager_list_jobs();
-}
-
-static json _rpc_op_load_or_unload(const json &args) {
-    // FIXME: handle Force and OverrideDisabled arguments
-    int (*manager_func)(const std::filesystem::path &);
-    if (args[0] == "load") {
-        manager_func = &manager_load_manifest;
-    } else if (args[0] == "unload") {
-        manager_func = &manager_unload_manifest;
+    auto maybe_job = mgr.at(label);
+    if (maybe_job) {
+        const auto job = *maybe_job;
+        job->kill(signame_or_num);
+        return {{"error", false}};
     } else {
-        throw std::logic_error("unsupported operation");
+        return {{"error", true}};
     }
+}
+
+static json _rpc_op_list(const json &, Manager &mgr) {
+    // FIXME: handle Label argument
+    return mgr.listJobs();
+}
+
+static json _rpc_op_load(const json &args, Manager &mgr) {
+    // FIXME: handle Force and OverrideDisabled arguments
     for (const auto &path : args[1]["Paths"]) {
         if (!std::filesystem::exists(path)) {
             return {{"error", true}};
@@ -68,62 +66,91 @@ static json _rpc_op_load_or_unload(const json &args) {
         if (std::filesystem::is_directory(path)) {
             using std::filesystem::directory_iterator;
             for (const auto &file: directory_iterator(path)) {
-                (*manager_func)(file.path());
+                mgr.loadManifest(file.path());
             }
         } else {
             std::filesystem::path p{path.get<std::string>()};
-            (*manager_func)(p);
+            mgr.loadManifest(p);
         }
     }
     return {{"error", false}};
 }
 
-static json _rpc_op_start(const json &args) {
+static json _rpc_op_unload(const json &args, Manager &mgr) {
+    // FIXME: handle Force and OverrideDisabled arguments
+    for (const auto &path : args[1]["Paths"]) {
+        if (!std::filesystem::exists(path)) {
+            return {{"error", true}};
+        }
+        if (std::filesystem::is_directory(path)) {
+            using std::filesystem::directory_iterator;
+            for (const auto &file: directory_iterator(path)) {
+                mgr.unloadJob(file.path());
+            }
+        } else {
+            std::filesystem::path p{path.get<std::string>()};
+            mgr.unloadJob(p);
+        }
+    }
+    return {{"error", false}};
+}
+
+static json _rpc_op_start(const json &args, Manager &mgr) {
     const std::string &label = args[1]["Label"];
-    auto job = manager_get_job_by_label(label);
-    job->run();
-    return {{"error", false}};
+    auto maybe_job = mgr.at(label);
+    if (maybe_job) {
+        const auto job = *maybe_job;
+        job->run();
+        return {{"error", false}};
+    } else {
+        return {{"error", true}};
+    }
 }
 
-static json _rpc_op_stop(const json &args) {
+static json _rpc_op_stop(const json &args, Manager &mgr) {
     const std::string &label = args[1]["Label"];
-    auto job = manager_get_job_by_label(label);
-    job->kill("SIGTERM");
+    auto maybe_job = mgr.at(label);
+    if (maybe_job) {
+        const auto job = *maybe_job;
+        job->kill("SIGTERM");
+        return {{"error", false}};
+    } else {
+        return {{"error", true}};
+    }
+}
+
+static json _rpc_op_remove(const json &args, Manager &mgr) {
+    const std::string &label = args[1]["Label"];
+    mgr.unloadJob(label);
     return {{"error", false}};
 }
 
-static json _rpc_op_remove(const json &args) {
-    const auto &label = args[1]["Label"];
-    manager_unload_by_label(label);
-    return {{"error", false}};
-}
-
-static json _rpc_op_submit(const json &args) {
+static json _rpc_op_submit(const json &args, Manager &mgr) {
     std::string path = ""; // TODO: maybe create a fake path? do we even need this?
-    manager_load_manifest(args[1], path);
+    mgr.loadManifest(args[1], path);
     return {{"error", false}};
 }
 
-static json _rpc_op_version(const json &) {
+static json _rpc_op_version(const json &, Manager &) {
     // FIXME get version number
     return {{"error", false},{"version", "relaunch version unknown"}};
 }
 
 // FIXME: needs a lot more error checking
-int rpc_dispatch(Channel &chan) {
-    static const std::unordered_map<std::string, json(*)(const json &j)> handlers = {
+int rpc_dispatch(Channel &chan, Manager &mgr) {
+    static const std::unordered_map<std::string, json(*)(const json &, Manager &)> handlers = {
             {"disable", _rpc_op_disable},
             {"enable", _rpc_op_enable},
             {"kill", _rpc_op_kill},
 
             {"list", _rpc_op_list},
-            {"load", _rpc_op_load_or_unload},
+            {"load", _rpc_op_load},
             {"remove", _rpc_op_remove},
             {"start", _rpc_op_start},
             {"submit", _rpc_op_submit},
 
             {"stop", _rpc_op_stop},
-            {"unload", _rpc_op_load_or_unload},
+            {"unload", _rpc_op_unload},
             {"version", _rpc_op_version},
     };
     chan.accept();
@@ -131,7 +158,7 @@ int rpc_dispatch(Channel &chan) {
         auto msg = chan.readMessage();
         auto method = msg.at(0).get<std::string>();
         auto funcptr = handlers.at(method);
-        json response = (*funcptr)(msg);
+        json response = (*funcptr)(msg, mgr);
         chan.writeMessage(response);
     } catch (const std::exception &exc) {
         log_error("dispatch failed: %s", exc.what());
